+++ src/PlanetsideSeaState.App/CensusStream/Models/FacilityControlPayload.cs
@@ -3,8 +3,8 @@
     public class FacilityControlPayload : PayloadBase, IEquatablePayload<FacilityControlPayload>
     {
         public int FacilityId { get; set; }
-        public int NewFactionId { get; set; }
-        public int OldFactionId { get; set; }
+        public short NewFactionId { get; set; }
+        public short OldFactionId { get; set; }
         public int DurationHeld { get; set; }
         public string OutfitId { get; set; }
