+++ src/PlanetsideSeaState.App/CensusStream/Models/PayloadBase.cs
@@ -5,8 +5,8 @@ namespace PlanetsideSeaState.App.CensusStream.Models
     public class PayloadBase
     {
         public string EventName { get; set; }
-        public int WorldId { get; set; }
-        public int? ZoneId { get; set; }
+        public short WorldId { get; set; }
+        public uint? ZoneId { get; set; }
         public DateTime Timestamp { get; set; }
     }
 }
