+++ src/PlanetsideSeaState.Graphing/Services/FacilityPopsGraph/Interfaces/IFacilityPopGraphServicesManager.cs
@@ -5,9 +5,9 @@ namespace PlanetsideSeaState.Graphing.Services.FacilityPopsGraph
 {
     public interface IFacilityPopGraphServicesManager : IStatefulHostedService
     {
-        FacilityPopGraphService GetService(int worldId, int zoneId);
+        FacilityPopGraphService GetService(short worldId, uint zoneId);
 
-        Task<bool> TryAddNewService(int worldId, int zoneId);
-        Task<bool> TryRemoveService(int worldId, int zoneId);
+        Task<bool> TryAddNewService(short worldId, uint zoneId);
+        Task<bool> TryRemoveService(short worldId, uint zoneId);
     }
 }
